Remove all the unused Salesforce classes and dependencies
